Use in-memory SQLite for testing
Use the in-memory sqlite db for testing. Changes have been made to
populate data for in-memory databases differently as there are now
no longer any migrations being used to create the schema when
in-memory SQLite is used (with exception of the explicit migration
test cases). The most notible case where the creation of the schema
differs from using the migrations, is the creation of the default
domain. Typically this has been done by migration 008. The restful
test cases now properly populate the default domain explicitly if
needed.

As part of this patch to ensure extensions received the correct
schema additions, the EXTENSION_NAME property is now used in the
test case to do migrations when not utilizing in-memory SQLite.

The model for ROLE table in assignment did not properly indicate
that the name column was "unique" (as would be the case using the
migrations). This unique column identifier is needed since the
migration no longer occurs (it becomes even more important that
the models match).

To support proper testing, the revoke tests in test_v3_auth now
provide thier own load_sample_data method and re-use the already
defined admin role based upon the super() call.

+++ keystone/tests/core.py
@@ -88,6 +88,8 @@ TMPDIR = _calc_tmpdir()
 
 CONF = config.CONF
 
+IN_MEM_DB_CONN_STRING = 'sqlite://'
+
 exception._FATAL_EXCEPTION_FORMAT_ERRORS = True
 os.makedirs(TMPDIR)
 atexit.register(shutil.rmtree, TMPDIR)
@@ -125,7 +127,7 @@ def _initialize_sql_session():
     # test cases.
     db_file = DEFAULT_TEST_DB_FILE
     db_options.set_defaults(
-        sql_connection='sqlite:///%s' % db_file,
+        sql_connection=IN_MEM_DB_CONN_STRING,
         sqlite_db=db_file)
 
 
@@ -164,19 +166,21 @@ def checkout_vendor(repo, rev):
     return revdir
 
 
-def setup_database():
-    db = dirs.tmp('test.db')
-    pristine = dirs.tmp('test.db.pristine')
+def setup_database(extensions=None):
+    if CONF.database.connection != IN_MEM_DB_CONN_STRING:
+        db = dirs.tmp('test.db')
+        pristine = dirs.tmp('test.db.pristine')
 
-    if os.path.exists(db):
-        os.unlink(db)
-    if not os.path.exists(pristine):
-        migration.db_sync(sql.get_engine(),
-                          migration_helpers.find_migrate_repo())
-        migration_helpers.sync_database_to_version(extension='revoke')
-        shutil.copyfile(db, pristine)
-    else:
-        shutil.copyfile(pristine, db)
+        if os.path.exists(db):
+            os.unlink(db)
+        if not os.path.exists(pristine):
+            migration.db_sync(sql.get_engine(),
+                              migration_helpers.find_migrate_repo())
+            for extension in (extensions or []):
+                migration_helpers.sync_database_to_version(extension=extension)
+            shutil.copyfile(db, pristine)
+        else:
+            shutil.copyfile(pristine, db)
 
 
 def teardown_database():
